% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-timePlotFormatting.R
\name{add_na_row}
\alias{add_na_row}
\title{Add NA Row}
\usage{
add_na_row(df)
}
\arguments{
\item{df}{(data.frame) data.frame}
}
\description{
Function to add a row with NA values at the end of a data.frame
}
