% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUploadModel.R
\name{addPackageVersionNo}
\alias{addPackageVersionNo}
\title{Add Package Version Number}
\usage{
addPackageVersionNo(txt)
}
\arguments{
\item{txt}{(character) text to which version number is appended}
}
\description{
Add Package Version Number
}
