% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{addXScale}
\alias{addXScale}
\title{Add X Scale}
\usage{
addXScale(p, object, oldXAxisData, deriv)
}
\arguments{
\item{p}{(ggplot) ggplot object}

\item{object}{model of class \code{\link{TemporalIso}}}

\item{oldXAxisData}{data.frame of time data from object}

\item{deriv}{character "1" for absolute values, "2" for first differences}
}
\description{
Add breaks and labels for x axis
}
