% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{adjustTimeColumn}
\alias{adjustTimeColumn}
\title{Adjust Time Column}
\usage{
adjustTimeColumn(objectTime, deriv)
}
\arguments{
\item{objectTime}{(numeric) vector, time element of object}

\item{deriv}{character "1" for absolute values, "2" for first differences}
}
\description{
Adjust Time Column
}
