% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-breakPointDetection.R
\name{breakPointDetectionUI}
\alias{breakPointDetectionUI}
\alias{breakPointDetectionServer}
\title{Break Point Detection UI}
\usage{
breakPointDetectionUI(id)

breakPointDetectionServer(id, plotData)
}
\arguments{
\item{id}{The module id}

\item{plotData}{The reactive plot data}
}
\description{
UI of the module

Server function of the module
}
