% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-cleanAndSplitData.R
\name{cleanAndSplitData}
\alias{cleanAndSplitData}
\title{Clean and Split Data}
\usage{
cleanAndSplitData(indVar, renewalRates, renewalRatesUnc)
}
\arguments{
\item{indVar}{(character) name of variable for individuals}

\item{renewalRates}{(matrix) matrix with renewal rates}

\item{renewalRatesUnc}{(matrix) matrix with renewal rates uncertainties}
}
\description{
Clean and Split Data
}
