% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-breakPoints.R
\name{compareWithHeuristic}
\alias{compareWithHeuristic}
\title{Compare models using heuristic}
\usage{
compareWithHeuristic(fit)
}
\arguments{
\item{fit}{A list of mcp model fits}
}
\value{
A loo model comparison object
}
\description{
The heuristic method is a model comparison method that combines the number of parameters and the elpd_diff value.
}
