% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-breakPoints.R
\name{compareWithLoo}
\alias{compareWithLoo}
\title{Compare models using loo}
\usage{
compareWithLoo(fit)
}
\arguments{
\item{fit}{A list of mcp model fits}
}
\value{
A loo model comparison object
}
\description{
Compare models using loo
}
