% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-detectBreakPoints.R
\name{detectBreakPoints}
\alias{detectBreakPoints}
\title{Detect break points in a time series}
\usage{
detectBreakPoints(
  df,
  x,
  y,
  burnin_len = 5000,
  chains = 3,
  iter = 3000,
  segments = exampleBreakPointSegments(),
  priors = exampleBreakPointPriors()
)
}
\arguments{
\item{df}{A data frame}

\item{x}{The x-axis column name}

\item{y}{The y-axis column name}
}
\description{
Detect break points in a time series
}
