% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{extractAllXAxisData}
\alias{extractAllXAxisData}
\title{Extract All X-Axis Data}
\usage{
extractAllXAxisData(models, allXAxisData = data.frame())
}
\arguments{
\item{models}{(list) list of models}

\item{allXAxisData}{(data.frame) empty data.frame, or a data.frame containing xAxisData, output
of \code{getXAxisData}}
}
\description{
Extract all x-axis data to draw x axis ticks and labels at all possible points in time present 
 in models
}
