% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-cleanAndSplitData.R
\name{extractIndividuals}
\alias{extractIndividuals}
\title{Extract Individuals}
\usage{
extractIndividuals(matrix, indVar)
}
\arguments{
\item{matrix}{(matrix) matrix of data}

\item{indVar}{(character) name if individuals column or empty if rownames}
}
\value{
(character) vector of individuals or row names with individuals
}
\description{
Extract Individuals
}
