% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUploadModel.R
\name{extractModel}
\alias{extractModel}
\title{Extract Model}
\usage{
extractModel(modelFromImport)
}
\arguments{
\item{modelFromImport}{imported model object}
}
\description{
Extract model fit and possibly model specification inputs from a previously exported
model object.
}
