% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{extractPlotData}
\alias{extractPlotData}
\title{Extract Plot Data}
\usage{
extractPlotData(object, prop = 0.8, deriv = "1")
}
\arguments{
\item{object}{model of class \code{\link{TemporalIso}}}

\item{prop}{double between 0 and 1: length of credibility interval. The default value is 80 percent.}

\item{deriv}{character "1" for absolute values, "2" for first differences}
}
\description{
Extract Plot Data
}
