% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-extractSavedModels.R
\name{extractSavedModels}
\alias{extractSavedModels}
\title{Extract Saved Models}
\usage{
extractSavedModels(upload)
}
\arguments{
\item{upload}{(list) list of saved model objects with entries "data", "inputs" and "model"}
}
\description{
Extract Saved Models
}
