% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-breakPointDetection.R
\name{formulasUI}
\alias{formulasUI}
\alias{formulasServer}
\title{Formulas UI}
\usage{
formulasUI(id)

formulasServer(id)
}
\arguments{
\item{id}{The module id}
}
\description{
Formulas UI

Formulas Server
}
