% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{getBreaks}
\alias{getBreaks}
\title{Get Breaks}
\usage{
getBreaks(time, deriv)
}
\arguments{
\item{time}{(numeric) time points of x axis}

\item{deriv}{character "1" for absolute values, "2" for first differences}
}
\description{
Get Breaks
}
