% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-matrixModule.R
\name{getCellChoices}
\alias{getCellChoices}
\title{Get cell choices}
\usage{
getCellChoices(nrow, ncol)
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of columns}
}
\value{
A named vector of cell choices
}
\description{
Get cell choices
}
