% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{getDefaultPlotRange}
\alias{getDefaultPlotRange}
\title{Get Default Plot Range}
\usage{
getDefaultPlotRange(savedModels, deriv = "1")
}
\arguments{
\item{savedModels}{list of models of class \code{\link{TemporalIso}}}

\item{deriv}{character "1" for absolute values, "2" for first differences}
}
\description{
Min/max values in x/y directions
}
