% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-helperFunctions.R
\name{getEntry}
\alias{getEntry}
\title{Get Entry}
\usage{
getEntry(modelList, entry)
}
\arguments{
\item{modelList}{(list) of saved model objects}

\item{entry}{(character) name of the list element, e.g. "fit"}
}
\description{
Get all list elements "entry" of a list of models
}
