% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-helperFunctions.R
\name{getExampleDataMatrix}
\alias{getExampleDataMatrix}
\title{Get Example Matrix of Data}
\usage{
getExampleDataMatrix(sd = FALSE)
}
\arguments{
\item{sd}{boolean - if true, get uncertainty renewal rates example data}
}
\description{
Get Example Matrix of Data
}
