% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plotSetXAxisLabels.R
\name{getLabel}
\alias{getLabel}
\title{Get Label}
\usage{
getLabel(xAxisData, deriv, hidePastedLabels = TRUE)
}
\arguments{
\item{xAxisData}{data.frame of time data of object, output of getXAxisData}

\item{deriv}{character "1" for absolute values, "2" for first differences}

\item{hidePastedLabels}{(logical) if TRUE then dont't show pasted labels}
}
\description{
Get Label
}
