% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateIntervals.R
\name{getSeed}
\alias{getSeed}
\title{Get Seed}
\usage{
getSeed(fixSeed, seedValue)
}
\arguments{
\item{fixSeed}{(logical) if TRUE use fixed ssed, if FALSE use random seed when fitting the model}

\item{seedValue}{(numeric) value for seed if \code{fixSeed == TRUE}}
}
\description{
Get Seed
}
