% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-module_modelSpecifications.R
\name{getTimeMax}
\alias{getTimeMax}
\title{Get Time Maximum}
\usage{
getTimeMax(mtrx, timeVars, default = 1)
}
\arguments{
\item{mtrx}{(matrix) data matrix}

\item{timeVars}{(character) column names of time variables}

\item{default}{(numeric) default result}
}
\value{
(numeric) maximal time
}
\description{
Get Time Maximum
}
