% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-module_modelSpecifications.R
\name{getTimeMin}
\alias{getTimeMin}
\title{Get Time Minimum}
\usage{
getTimeMin(mtrx, timeVars, default = 0)
}
\arguments{
\item{mtrx}{(matrix) data matrix}

\item{timeVars}{(character) column names of time variables}

\item{default}{(numeric) default result}
}
\value{
(numeric) minimal time
}
\description{
Get Time Minimum
}
