% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift.R
\name{getUserDefinedEst}
\alias{getUserDefinedEst}
\title{Calculate estimates for user defined intervals}
\usage{
getUserDefinedEst(object, minim = 0, maxim = 1, type = "1")
}
\arguments{
\item{object}{model of class \code{\link{TemporalIso}}}

\item{minim}{numeric lower interval point}

\item{maxim}{numeric upper interval point}

\item{type}{character "1" for absolute values, "2" for total turnover}
}
\value{
a data.frame containing mean and sd within the interval
}
\description{
Calculate estimates for user defined intervals
}
