% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{getXAxisData}
\alias{getXAxisData}
\title{Get X-Axis Data}
\usage{
getXAxisData(object, oldXAxisData = data.frame())
}
\arguments{
\item{object}{model of class \code{\link{TemporalIso}}}

\item{oldXAxisData}{data.frame, output from previous run of this function}
}
\description{
Get X-Axis Data
}
