% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-breakPointDetection.R
\name{infoButtonUI}
\alias{infoButtonUI}
\alias{infoButtonServer}
\title{Info Button UI}
\usage{
infoButtonUI(id, label = "Description")

infoButtonServer(id, title = "Description", text, link = NULL)
}
\arguments{
\item{id}{The module id}

\item{label}{The label of the button}

\item{title}{The title of the modal}

\item{text}{The text to display in the modal}

\item{link}{The link to the documentation}
}
\description{
Info Button UI

Info Button Server
}
