% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{joinTimeForDerivation}
\alias{joinTimeForDerivation}
\title{Get Breaks for Derivation}
\usage{
joinTimeForDerivation(time)
}
\arguments{
\item{time}{(numeric) time points of x axis}
}
\description{
Get the time for the x axis of the plot in case of showing the derivation
}
