% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-matrixModule.R
\name{matrixUI}
\alias{matrixUI}
\alias{matrixServer}
\title{Matrix Module}
\usage{
matrixUI(
  id,
  title = "Matrix",
  maxRows = 10,
  maxColumns = 5,
  defaultCellContent = "",
  exampleLabel = "Example Matrix"
)

matrixServer(id, exampleFunction, validateCellFunction = NULL, ...)
}
\arguments{
\item{id}{namespace id}

\item{maxRows}{maximum number of rows}

\item{maxColumns}{maximum number of columns}

\item{defaultCellContent}{default cell content}

\item{exampleLabel}{example label}

\item{exampleFunction}{example function}

\item{...}{additional arguments to example function}
}
\description{
Matrix Module

Matrix Server
}
