% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-breakPointDetection.R
\name{mcpCompareModelsUI}
\alias{mcpCompareModelsUI}
\alias{mcpCompareModelsServer}
\title{MCP Compare Models UI}
\usage{
mcpCompareModelsUI(id)

mcpCompareModelsServer(id, formulasAndPriors, mcpData, mcpFitList)
}
\arguments{
\item{id}{The module id}

\item{formulasAndPriors}{The reactive formulas and priors}

\item{mcpData}{The reactive mcp data}

\item{mcpFitList}{The reactive mcp fit list}
}
\description{
MCP Compare Models UI

MCP Compare Models Server
}
