% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-breakPointDetection.R
\name{mcpFormulasUI}
\alias{mcpFormulasUI}
\alias{mcpFormulasServer}
\title{MCP Formulas UI}
\usage{
mcpFormulasUI(id)

mcpFormulasServer(id)
}
\arguments{
\item{id}{The module id}
}
\description{
MCP Formulas UI

MCP Formulas Server
}
