% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-breakPointDetection.R
\name{mcpOutUI}
\alias{mcpOutUI}
\alias{mcpOutServer}
\title{MCP Output UI}
\usage{
mcpOutUI(id, title, outFUN = verbatimTextOutput, showWidth = FALSE)

mcpOutServer(
  id,
  formulasAndPriors,
  mcpData,
  mcpFitList,
  mcpModel,
  outFUN,
  renderFUN = renderPrint
)
}
\arguments{
\item{id}{The module id}

\item{title}{The title of the output}

\item{outFUN}{The output function}

\item{showWidth}{Show the width slider}

\item{formulasAndPriors}{The reactive formulas and priors}

\item{mcpData}{The reactive mcp data}

\item{mcpFitList}{The reactive mcp fit list}

\item{mcpModel}{The reactive mcp model}

\item{renderFUN}{The render function}
}
\description{
MCP Output UI

MCP Output Server
}
