% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-breakPointDetection.R
\name{mcpShowSingleModelUI}
\alias{mcpShowSingleModelUI}
\alias{mcpShowSingleModelServer}
\title{MCP Show Single Model UI}
\usage{
mcpShowSingleModelUI(id)

mcpShowSingleModelServer(id, mcpData, formulasAndPriors, mcpFitList)
}
\arguments{
\item{id}{The module id}

\item{mcpData}{The reactive mcp data}

\item{formulasAndPriors}{The reactive formulas and priors}

\item{mcpFitList}{The reactive mcp fit list}
}
\description{
MCP Show Single Model UI

MCP Show Single Model Server
}
