% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-module_modelSpecifications.R
\name{modelSpecificationsServer}
\alias{modelSpecificationsServer}
\title{server funtion of modelSpecifications module}
\usage{
modelSpecificationsServer(id, dataMatrix, uploadedModelSpecInputs)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{dataMatrix}{(reactive) shinyMatrix matrixInput, see \code{\link{estimateIntervals}}: 
specifying the renewal rates of different probes for each time interval.
The renewalRates should be between 0 and 100 (percentages). The dataframe should include
a column specifying a time-index (e.g. 1, 2, 3, ...)
as well as columns for the different bones. The renewal rates should contain the times of origin,
containing 100.}

\item{uploadedModelSpecInputs}{(reactive) modelSpecifications for the current fit}
}
\description{
server funtion of modelSpecifications module
}
