% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-module_modelSpecifications.R
\name{modelSpecificationsUI}
\alias{modelSpecificationsUI}
\title{ui function of modelSpecifications module}
\usage{
modelSpecificationsUI(id, title)
}
\arguments{
\item{id}{namespace}

\item{title}{title in tab}
}
\description{
ui function of modelSpecifications module
}
