% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pasteLabelForDerivation}
\alias{pasteLabelForDerivation}
\title{Paste Label for Derivation Plot}
\usage{
pasteLabelForDerivation(axesLabel, hidePastedLabels)
}
\arguments{
\item{axesLabel}{(character) output of \code{\link{getLabel}}}

\item{hidePastedLabels}{(logical) if TRUE then dont't show pasted labels}
}
\description{
Get the labels for the x axis of the plot in case of showing the derivation
}
