% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTime}
\alias{plotTime}
\title{Plot of credibility intervals for each time interval, plotted as timeseries}
\usage{
plotTime(
  object,
  prop = 0.8,
  plotShifts = FALSE,
  yLim = c(0, 1),
  xLim = c(0, 1),
  deriv = "1",
  color = NULL,
  alphaL = 0.9,
  alphaU = 0.1,
  sizeTextY = 12,
  sizeTextX = 12,
  sizeAxisX = 12,
  sizeAxisY = 12,
  secAxis = FALSE,
  xAxisLabel = "Time",
  yAxisLabel = "Estimate",
  extendLabels = FALSE,
  pointStyle = config()[["defaultPointStyle"]],
  ...
)
}
\arguments{
\item{object}{model of class \code{\link{TemporalIso}}}

\item{prop}{double between 0 and 1: length of credibility interval. The default value is 80 percent.}

\item{plotShifts}{boolean if shifts shall be marked or not. Defaults to False.}

\item{yLim}{numeric vector of length 2: range of y axis}

\item{xLim}{numeric vector of length 2: range of x axis}

\item{deriv}{character "1" for absolute values, "2" for first differences}

\item{color}{color of line and uncertainty region (RGB)}

\item{alphaL}{opacity line}

\item{alphaU}{opacity uncertainty region}

\item{sizeTextY}{size plot y label}

\item{sizeTextX}{size text x label}

\item{sizeAxisX}{size text x axis label}

\item{sizeAxisY}{size text y axis label}

\item{secAxis}{boolean secondary axis on right side?}

\item{xAxisLabel}{character label x-axis}

\item{yAxisLabel}{character label y-axis}

\item{extendLabels}{boolean if TRUE, extend the labels of the x-axis to the x-axis limits. 
If FALSE, the range of the data defines the range of x-axis labels.}

\item{pointStyle}{(list) named list with style definitions, or output of \code{plotPointsServer}}

\item{...}{arguments handed to \code{\link{getShiftTime}}}
}
\value{
a \link[ggplot2]{ggplot} object.
}
\description{
The function plots the credibility intervals for each
time interval and returns a ggplot object, which is further customizable.
}
