% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-breakPoints.R
\name{setFormulasAndPriors}
\alias{setFormulasAndPriors}
\title{Set formulas and priors}
\usage{
setFormulasAndPriors(splittedComb)
}
\arguments{
\item{splittedComb}{A list of two matrices containing the segmented formulas and priors}
}
\value{
A list of lists containing the segmented formulas and priors
}
\description{
Set formulas and priors
}
