% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-helperFunctions.R
\name{setVarsForUncMatrix}
\alias{setVarsForUncMatrix}
\title{Set Vars For Unc Matrix}
\usage{
setVarsForUncMatrix(timeVars, indVar, renewalRates, renewalRatesUnc = NULL)
}
\arguments{
\item{timeVars}{A character string specifying the name of the column indicating the time.}

\item{indVar}{(character) name of variable for individuals}

\item{renewalRates}{(matrix) matrix with renewal rates}

\item{renewalRatesUnc}{(matrix) matrix with renewal rates uncertainties}
}
\description{
Set Vars For Unc Matrix
}
