% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMESPACE.R
\name{shinyModule}
\alias{shinyModule}
\title{Server and UI Functions for Shiny Module}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session}

\item{id}{namespace id}

\item{title}{title of tab in tabset panel}
}
\description{
Server and UI Functions for Shiny Module
}
