% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-breakPoints.R
\name{splitComb}
\alias{splitComb}
\title{Split the comb matrix into two matrices}
\usage{
splitComb(comb)
}
\arguments{
\item{comb}{A matrix containing the segmented formulas and priors}
}
\value{
A list of two matrices
}
\description{
Split the comb matrix into two matrices
}
