% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-timePlotFormatting.R
\name{timePlotFormattingServer}
\alias{timePlotFormattingServer}
\title{Server function of time plot formatting}
\usage{
timePlotFormattingServer(id, savedModels)
}
\arguments{
\item{id}{namespace id}

\item{savedModels}{list of models of class \code{\link{TemporalIso}}}
}
\description{
Backend for plot formatting module
}
