% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-timePlotFormatting.R
\name{timePlotFormattingUI}
\alias{timePlotFormattingUI}
\title{UI function of time plot formatting}
\usage{
timePlotFormattingUI(id)
}
\arguments{
\item{id}{module id}
}
\value{
actionButton
}
\description{
UI function of time plot formatting
}
