% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-helperFunctions.R
\name{updateMatrixNamesInput}
\alias{updateMatrixNamesInput}
\title{Update only matrix row/colnames}
\usage{
updateMatrixNamesInput(session, inputId, value, value2)
}
\description{
Used for renewal uncertainties
}
