% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{updateTime}
\alias{updateTime}
\title{Extract Plot Data}
\usage{
updateTime(plotData, time, deriv = "1")
}
\arguments{
\item{plotData}{(data.frame) plot data}

\item{time}{(numeric) time entry of object (model of class \code{\link{TemporalIso}})}

\item{deriv}{character "1" for absolute values, "2" for first differences}
}
\description{
Extract Plot Data
}
