% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downAndUploadModel.R
\name{uploadModelUI}
\alias{uploadModelUI}
\alias{uploadModel}
\title{Upload model module}
\usage{
uploadModelUI(id, label)

uploadModel(
  input,
  output,
  session,
  savedModels,
  uploadedNotes,
  fit,
  uploadedModelSpecInputs,
  uploadedDataMatrix,
  uploadedDataMatrixSD,
  uploadedIsotope
)
}
\arguments{
\item{id}{id of module}

\item{label}{label of module}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{savedModels}{(reactive) list of models of class \code{\link{TemporalIso}}}

\item{uploadedNotes}{(reactive) variable that stores content of README.txt}

\item{fit}{(reactive) model of class \code{\link{TemporalIso}} that is currently displayed}

\item{uploadedModelSpecInputs}{(reactive) modelSpecifications for the current fit}

\item{uploadedDataMatrix}{(reactive) shinyMatrix matrixInput}

\item{uploadedDataMatrixSD}{(reactive) shinyMatrix matrixInput}

\item{uploadedIsotope}{(reactive) shinyMatrix matrixInput}
}
\description{
UI function to upload a zip file with notes and a list of models

Backend for upload model module
}
