% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-matrixModule.R
\name{validateFormula}
\alias{validateFormula}
\title{Check formula syntax}
\usage{
validateFormula(formula)
}
\arguments{
\item{formula}{A formula to check}
}
\value{
A formula if valid, otherwise an error
}
\description{
Check if the formula syntax is valid
}
