% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-breakPoints.R
\name{cleanComb}
\alias{cleanComb}
\title{Clean the comb matrix}
\usage{
cleanComb(comb)
}
\arguments{
\item{comb}{A matrix containing the segmented formulas and priors}
}
\value{
A cleaned matrix
}
\description{
Clean the comb matrix
}
