% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-timePlotFormatting.R
\name{timePlotFormattingServer}
\alias{timePlotFormattingServer}
\title{Server function of time plot formatting}
\usage{
timePlotFormattingServer(id, fit, savedModels)
}
\arguments{
\item{id}{namespace id}
}
\description{
Backend for plot formatting module
}
