% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-helperFunctions.R
\name{updateMatrixNamesInput}
\alias{updateMatrixNamesInput}
\title{Update only matrix row/colnames}
\usage{
updateMatrixNamesInput(session, inputId, value, value2)
}
\arguments{
\item{session}{session}

\item{inputId}{input Id}

\item{value}{matrix of renewal rates}

\item{value2}{matrix of uncertainties of renewal rates}
}
\description{
Used for renewal uncertainties
}
